# Improve performance of `ModelBuilder.add_builder` method