--- a/newton/_src/sim/builder.py
+++ b/newton/_src/sim/builder.py
@@ -361,7 +361,7 @@ class ModelBuilder:
         self.shape_group = []
 
         # filtering to ignore certain collision pairs
-        self.shape_collision_filter_pairs = set()
+        self.shape_collision_filter_pairs: list[tuple[int, int]] = []
 
         # springs
         self.spring_indices = []
@@ -636,9 +636,10 @@ class ModelBuilder:
                 Defaults to (5.0, 5.0, 0.0).
         """
         offsets = self._compute_replicate_offsets(num_copies, spacing)
+        xform = wp.transform_identity()
         for i in range(num_copies):
-            with wp.ScopedTimer("add_builder"):
-                self.add_builder(builder, xform=wp.transform(offsets[i], wp.quat_identity()))
+            xform.p = offsets[i]
+            self.add_builder(builder, xform=xform)
 
     def add_articulation(self, key: str | None = None):
         # an articulation is a set of contiguous bodies bodies from articulation_start[i] to articulation_start[i+1]
@@ -964,7 +965,8 @@ class ModelBuilder:
 
         # dispatches two transform multiplies to the native implementation
         def transform_mul(a, b):
-            out = wp.transformf()
+            # out = wp.transformf()
+            out = wp.transform.from_buffer(np.empty(7, dtype=np.float32))
             transform_mul_cfunc(a, b, ctypes.byref(out))
             return out
 
@@ -972,7 +974,7 @@ class ModelBuilder:
         if builder.particle_count:
             self.particle_max_velocity = builder.particle_max_velocity
             if xform is not None:
-                pos_offset = wp.transform_get_translation(xform)
+                pos_offset = xform.p
             else:
                 pos_offset = np.zeros(3)
             self.particle_q.extend((np.array(builder.particle_q) + pos_offset).tolist())
@@ -1007,7 +1009,7 @@ class ModelBuilder:
                 self.shape_body.append(-1)
                 # apply offset transform to root bodies
                 if xform is not None:
-                    self.shape_transform.append(transform_mul(xform, wp.transform(*builder.shape_transform[s])))
+                    self.shape_transform.append(transform_mul(xform, builder.shape_transform[s]))
                 else:
                     self.shape_transform.append(builder.shape_transform[s])
 
@@ -1034,7 +1036,7 @@ class ModelBuilder:
                         self.joint_q[qi : qi + 3] = tf.p
                         self.joint_q[qi + 3 : qi + 7] = tf.q
                     elif builder.joint_parent[i] == -1:
-                        self.joint_X_p[start_X_p + i] = transform_mul(xform, wp.transform(*builder.joint_X_p[i]))
+                        self.joint_X_p[start_X_p + i] = transform_mul(xform, builder.joint_X_p[i])
 
             # offset the indices
             self.articulation_start.extend([a + self.joint_count for a in builder.articulation_start])
@@ -1044,19 +1046,20 @@ class ModelBuilder:
             self.joint_q_start.extend([c + self.joint_coord_count for c in builder.joint_q_start])
             self.joint_qd_start.extend([c + self.joint_dof_count for c in builder.joint_qd_start])
 
-        for i in range(builder.body_count):
-            if xform is not None:
-                self.body_q.append(transform_mul(xform, wp.transform(*builder.body_q[i])))
-            else:
-                self.body_q.append(builder.body_q[i])
+        if xform is not None:
+            for i in range(builder.body_count):
+                self.body_q.append(transform_mul(xform, builder.body_q[i]))
+        else:
+            self.body_q.extend(builder.body_q)
 
         # Copy collision groups without modification
         self.shape_collision_group.extend(builder.shape_collision_group)
 
         # Copy collision filter pairs with offset
         shape_count_offset = self.shape_count
-        for i, j in builder.shape_collision_filter_pairs:
-            self.shape_collision_filter_pairs.add((i + shape_count_offset, j + shape_count_offset))
+        self.shape_collision_filter_pairs.extend(
+            [(i + shape_count_offset, j + shape_count_offset) for i, j in builder.shape_collision_filter_pairs]
+        )
 
         # Handle environment group assignments
         # For particles
@@ -1357,7 +1360,7 @@ class ModelBuilder:
                     a, b = parent_shape, child_shape
                     if a > b:
                         a, b = b, a
-                    self.shape_collision_filter_pairs.add((a, b))
+                    self.shape_collision_filter_pairs.append((a, b))
 
         return self.joint_count - 1
 
@@ -2222,8 +2225,8 @@ class ModelBuilder:
             new_id = len(self.body_key)
             body_remap[body["original_id"]] = new_id
             self.body_key.append(body["key"])
-            self.body_q.append(list(body["q"]))
-            self.body_qd.append(list(body["qd"]))
+            self.body_q.append(body["q"])
+            self.body_qd.append(body["qd"])
             m = body["mass"]
             inertia = body["inertia"]
             self.body_mass.append(m)
@@ -2303,8 +2306,8 @@ class ModelBuilder:
             self.joint_qd.extend(joint["qd"])
             self.joint_armature.extend(joint["armature"])
             self.joint_enabled.append(joint["enabled"])
-            self.joint_X_p.append(list(joint["parent_xform"]))
-            self.joint_X_c.append(list(joint["child_xform"]))
+            self.joint_X_p.append(joint["parent_xform"])
+            self.joint_X_c.append(joint["child_xform"])
             self.joint_dof_dim.append(joint["axis_dim"])
             # Rebuild joint group - use original group if it exists
             if original_joint_group and joint["original_id"] < len(original_joint_group):
@@ -2410,7 +2413,7 @@ class ModelBuilder:
         if body in self.body_shapes:
             # no contacts between shapes of the same body
             for same_body_shape in self.body_shapes[body]:
-                self.shape_collision_filter_pairs.add((same_body_shape, shape))
+                self.shape_collision_filter_pairs.append((same_body_shape, shape))
             self.body_shapes[body].append(shape)
         else:
             self.body_shapes[body] = [shape]
@@ -2435,7 +2438,7 @@ class ModelBuilder:
             for parent_body in self.joint_parents[body]:
                 if parent_body > -1:
                     for parent_shape in self.body_shapes[parent_body]:
-                        self.shape_collision_filter_pairs.add((parent_shape, shape))
+                        self.shape_collision_filter_pairs.append((parent_shape, shape))
 
         if not is_static and cfg.density > 0.0:
             (m, c, I) = compute_shape_inertia(type, scale, src, cfg.density, cfg.is_solid, cfg.thickness)
@@ -4425,7 +4428,7 @@ class ModelBuilder:
             - Sets `model.shape_contact_pair_count` to the number of contact pairs found.
         """
         # Copy the set of filtered-out shape pairs to avoid modifying the original
-        filters = copy.copy(self.shape_collision_filter_pairs)
+        filters = set(self.shape_collision_filter_pairs)
         contact_pairs = []
 
         # Sort shapes by env group in case they are not sorted, keep only colliding shapes