--- a/newton/_src/utils/import_urdf.py
+++ b/newton/_src/utils/import_urdf.py
@@ -532,12 +532,12 @@ def parse_urdf(
 
     for i in range(start_shape_count, end_shape_count):
         for j in visual_shapes:
-            builder.shape_collision_filter_pairs.add((i, j))
+            builder.shape_collision_filter_pairs.append((i, j))
 
     if not enable_self_collisions:
         for i in range(start_shape_count, end_shape_count):
             for j in range(i + 1, end_shape_count):
-                builder.shape_collision_filter_pairs.add((i, j))
+                builder.shape_collision_filter_pairs.append((i, j))
 
     if collapse_fixed_joints:
         builder.collapse_fixed_joints()