<!--
Thank you for contributing to Newton!

Please fill the relevant sections.

Checkboxes can also be marked after you submit the PR.
-->

## Description
Improve performance of `add_builder`, `replicate`, and `find_shape_contact_pairs` methods in `ModelBuilder`.
Add unit tests for `add_builder` method.
Addresses https://github.com/newton-physics/newton/issues/55.

### Time Measurements

| Environment (8192 envs) | Function                 | Before  | After |
|--------------------------|--------------------------|---------|-------|
| Allegro hand             | ModelBuilder.replicate   | 28.3 s  | 2.8 s |
|                          | ModelBuilder.finalize    | 3.7 s   | 2.9 s |
| G1 humanoid              | ModelBuilder.replicate   | 47.6 s  | 5.1 s |
|                          | ModelBuilder.finalize    | 4.2 s   | 5.1 s |
| Cartpole                 | ModelBuilder.replicate   | 5.8 s   | 0.8 s |
|                          | ModelBuilder.finalize    | 0.4 s   | 0.3 s |


## Newton Migration Guide

Please ensure the migration guide for **warp.sim** users is up-to-date with the changes made in this PR.

- [ ] The migration guide in ``docs/migration.rst`` is up-to date

## Before your PR is "Ready for review"

- [ ] Necessary tests have been added and new examples are tested (see `newton/tests/test_examples.py`)
- [ ] Documentation is up-to-date
- [ ] Code passes formatting and linting checks with `pre-commit run -a`


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Refactor**
  * Collision-filter pairs now preserve insertion order and may include duplicates, changing collision-filtering behavior across imports and model composition.
  * Model composition and replication are more efficient—transform reuse and reduced allocations improve merge/replication performance and memory use.

* **Tests**
  * Added test coverage validating composition of multiple builders and transform propagation.

* **Documentation**
  * Example updated to reflect new collision-filter handling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->